Drop Elasticsearch5QueryCompiler and make Elasticsearch6QueryCompiler the base implementation
